refactor(specs): coding style and move folders
Add better style in all index spec/admin/controllers
Move helpers folder into support folder and add configs folder

close issue #35
