+++ app/views/admin/daily_questions/index.html.haml
@@ -1,1 +1,1 @@
-= @daily_questions.map {|d| d.scheduled_at }
+= @daily_questions.map(&:scheduled_at)
