+++ app/views/admin/documentations/index.html.haml
@@ -1,1 +1,1 @@
-= @documentations.map {|d| d.name }
+= @documentations.map(&:name)
