+++ app/views/admin/questions/index.html.haml
@@ -1,1 +1,1 @@
-= @questions.map {|q| q.body }
+= @questions.map(&:body)
