+++ spec/controllers/admin/daily_questions_controller_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 
 RSpec.describe Admin::DailyQuestionsController do
-  login_admin
+  before(:each) { login_admin }
 
   describe 'GET index' do
     it 'returns http success' do
