+++ spec/controllers/home_controller_spec.rb
@@ -1,12 +1,10 @@
 require 'rails_helper'
 
-RSpec.describe HomeController, :type => :controller do
-
-  describe "GET index" do
-    it "returns http success" do
+RSpec.describe HomeController do
+  describe 'GET index' do
+    it 'returns http success' do
       get :index
       expect(response).to have_http_status(:success)
     end
   end
-
 end
