--- spec/helpers/controllers_spec_helper.rb
@@ -1,19 +1,0 @@
-module ControllerHelpers
-  def login_admin
-    before(:each) do
-      @request.env['devise.mapping'] = Devise.mappings[:admin]
-      admin = FactoryGirl.create(:admin)
-      admin.confirm
-      sign_in :user, admin # sign_in(scope, resource)
-    end
-  end
-
-  def login_user
-    before(:each) do
-      @request.env['devise.mapping'] = Devise.mappings[:user]
-      user = FactoryGirl.create(:user)
-      user.confirm
-      sign_in user
-    end
-  end
-end
