+++ spec/support/helpers/controllers_spec_helper.rb
@@ -1,0 +1,15 @@
+module ControllerHelpers
+  def login_admin
+    @request.env['devise.mapping'] = Devise.mappings[:admin]
+    admin = FactoryGirl.create(:admin)
+    admin.confirm
+    sign_in :user, admin # sign_in(scope, resource)
+  end
+
+  def login_user
+    @request.env['devise.mapping'] = Devise.mappings[:user]
+    user = FactoryGirl.create(:user)
+    user.confirm
+    sign_in user
+  end
+end
